feat: use fnv-1a hashing algorithm
Cloud variables are now fulling functional (at least on PC and Wii U)

+++ source/main.cpp
@@ -18,7 +18,7 @@
 // ^ for debug purposes
 
 #ifdef ENABLE_CLOUDVARS
-size_t projectHash;
+uint64_t projectHash;
 std::unique_ptr<MistConnection> cloudConnection = nullptr;
 #endif
